+++ src/line.rs
@@ -1,0 +1,33 @@
+use crate::tree;
+use std::{ops::RangeInclusive, time::Duration};
+
+#[derive(Clone)]
+pub struct Options {
+    /// If set, specify all levels that should be shown. Otherwise all available levels are shown.
+    ///
+    /// This is useful to filter out high-noise lower level progress items in the tree.
+    pub level_filter: Option<RangeInclusive<tree::Level>>,
+
+    /// If set, progress will only actually be shown after the given duration.
+    ///
+    /// This option can be useful to not enforce progress for short actions, causing it to flicker.
+    /// Please note that this won't affect display of messages, which are simply logged.
+    pub initial_delay: Option<Duration>,
+
+    /// The amount of frames to draw per second. If below 1.0, it determines the amount of seconds between the frame.
+    ///
+    /// *e.g.* 1.0/4.0 is one frame every 4 seconds.
+    pub frames_per_second: f32,
+
+    /// If true (default: false), we will keep waiting for progress even after we encountered an empty list of drawable progress items.
+    ///
+    /// Please note that you should add at least one item to the `prodash::Tree` before launching the application or else
+    /// risk a race causing nothing to be rendered at all.
+    pub keep_running_if_progress_is_empty: bool,
+}
+
+pub struct JoinHandle(std::thread::JoinHandle<()>);
+
+pub fn render(out: impl std::io::Write, progress: tree::Root, config: Options) -> JoinHandle {
+    unimplemented!("hello")
+}